Rename project to p4pillon
