+++ p4pillon/nthandlers.py
@@ -9,9 +9,9 @@ from collections.abc import Callable
 from p4p import Value
 from p4p.server import ServerOperation
 
-from p4p_ext.composite_handler import AbortHandlerException
-from p4p_ext.server.raw import Handler, SharedPV
-from p4p_ext.utils import overwrite_unmarked
+from p4pillon.composite_handler import AbortHandlerException
+from p4pillon.server.raw import Handler, SharedPV
+from p4pillon.utils import overwrite_unmarked
 
 from .rules import (
     AlarmRule,
