+++ p4pillon/server/thread.py
@@ -3,12 +3,12 @@ Monkey patch in required changes to Handlers and SharedPVs
 """
 
 ####
-# First override the base class of p4p_ext.server.thread.SharedPV with
-# p4p_ext.server.raw.SharedPV. This requires us to perform the imports
+# First override the base class of p4pillon.server.thread.SharedPV with
+# p4pillon.server.raw.SharedPV. This requires us to perform the imports
 # in a very specific order, which means overriding Linter checks
 import p4p.server.raw
 
-from p4p_ext.server.raw import SharedPV as _SharedPV
+from p4pillon.server.raw import SharedPV as _SharedPV
 
 p4p.server.raw.SharedPV = _SharedPV
 
@@ -19,6 +19,6 @@ from p4p.server.thread import Handler, SharedPV  # noqa: E402, F401,
 # Monkey patching the Handler is a simpler operation as it's a straight
 # substitution with our new version.
 # pylint: disable=ungrouped-imports
-from p4p_ext.server.raw import Handler as _Handler  # noqa: E402
+from p4pillon.server.raw import Handler as _Handler  # noqa: E402
 
 Handler = _Handler  # noqa: F811
