+++ p4pillon/thread/config_reader.py
@@ -7,10 +7,10 @@ from typing import Any
 
 import yaml
 
-from p4p_ext.definitions import PVTypes
-from p4p_ext.pvrecipe import BasePVRecipe
-from p4p_ext.server.simpleserver import BaseSimpleServer
-from p4p_ext.thread.pvrecipe import PVEnumRecipe, PVScalarArrayRecipe, PVScalarRecipe
+from p4pillon.definitions import PVTypes
+from p4pillon.pvrecipe import BasePVRecipe
+from p4pillon.server.simpleserver import BaseSimpleServer
+from p4pillon.thread.pvrecipe import PVEnumRecipe, PVScalarArrayRecipe, PVScalarRecipe
 
 logger = logging.getLogger(__name__)
