+++ p4pillon/thread/pvrecipe.py
@@ -2,10 +2,10 @@
 
 from __future__ import annotations
 
-from p4p_ext.definitions import PVTypes
-from p4p_ext.pvrecipe import BasePVRecipe
-from p4p_ext.pvrecipe import PVScalarRecipe as _PVScalarRecipe
-from p4p_ext.server.thread import SharedPV
+from p4pillon.definitions import PVTypes
+from p4pillon.pvrecipe import BasePVRecipe
+from p4pillon.pvrecipe import PVScalarRecipe as _PVScalarRecipe
+from p4pillon.server.thread import SharedPV
 
 
 class PVScalarRecipe(_PVScalarRecipe):
