+++ p4pillon/thread/server.py
@@ -9,9 +9,9 @@ import logging
 from p4p.client.thread import Context
 from p4p.server import Server, StaticProvider
 
-from p4p_ext.pvrecipe import BasePVRecipe
-from p4p_ext.server.simpleserver import BaseSimpleServer
-from p4p_ext.server.thread import SharedPV
+from p4pillon.pvrecipe import BasePVRecipe
+from p4pillon.server.simpleserver import BaseSimpleServer
+from p4pillon.server.thread import SharedPV
 
 logger = logging.getLogger(__name__)
